Move drive cache code to separate file
