+++ src/routes/cloud/+page.svelte
@@ -12,6 +12,7 @@
   import { exportAndUploadVolumesToDrive } from '$lib/util/cloud';
   import driveStore, { setDriveToken, clearDriveToken, fetchAllDriveData } from '$lib/util/drive-store';
   import { DriveErrorType } from '$lib/util/api-helpers';
+  import driveCache, { getFileId, listFilesInFolder } from '$lib/util/drive-cache';
   import { writable } from 'svelte/store';
 
   let accessToken = '';
@@ -525,11 +526,13 @@
       driveCache.clear();
       
       // Find the reader folder
-      const readerFolders = await driveCache.fetchFiles({
+      const readerFolders = await listFilesInFolder(null, {
         mimeTypes: [FOLDER_MIME_TYPE],
         name: READER_FOLDER,
         fields: 'files(id, name, mimeType, parents)',
-        context: 'checking for reader folder'
+        context: 'checking for reader folder',
+        forceRefresh: true,
+        handleError: false
       });
 
       if (readerFolders.length === 0) {
@@ -1016,216 +1019,6 @@
   }
 
   /**
-   * Simple virtual Google Drive cache to minimize API calls
-   * This stores file metadata in a structure similar to Google Drive
-   */
-  const driveCache = {
-    // Flag to indicate if the cache is initialized
-    initialized: false,
-    
-    // Flag to indicate if the cache is stale and needs refreshing
-    isStale: false,
-    
-    // Store files by ID for quick lookup
-    filesById: {},
-    
-    // Store files by parent ID for folder navigation
-    filesByParent: {},
-    
-    // Store files by name and parent for quick lookups
-    filesByNameAndParent: {},
-    
-    // Add a file to the cache
-    addFile(file) {
-      if (!file || !file.id) return;
-      
-      // Store by ID
-      this.filesById[file.id] = file;
-      
-      // Store by parent
-      if (file.parents && file.parents.length > 0) {
-        const parentId = file.parents[0];
-        if (!this.filesByParent[parentId]) {
-          this.filesByParent[parentId] = [];
-        }
-        
-        // Check if file already exists in this parent's array
-        const existingIndex = this.filesByParent[parentId].findIndex(f => f.id === file.id);
-        if (existingIndex >= 0) {
-          // Update existing entry
-          this.filesByParent[parentId][existingIndex] = file;
-        } else {
-          // Add new entry
-          this.filesByParent[parentId].push(file);
-        }
-        
-        // Store by name and parent for quick lookups
-        const key = `${parentId}:${file.name}`;
-        this.filesByNameAndParent[key] = file;
-      }
-    },
-    
-    // Remove a file from the cache
-    removeFile(fileId) {
-      if (!fileId || !this.filesById[fileId]) return;
-      
-      const file = this.filesById[fileId];
-      
-      // Remove from filesById
-      delete this.filesById[fileId];
-      
-      // Remove from filesByParent
-      if (file.parents && file.parents.length > 0) {
-        const parentId = file.parents[0];
-        if (this.filesByParent[parentId]) {
-          this.filesByParent[parentId] = this.filesByParent[parentId].filter(f => f.id !== fileId);
-          
-          // Remove from filesByNameAndParent
-          const key = `${parentId}:${file.name}`;
-          delete this.filesByNameAndParent[key];
-        }
-      }
-    },
-    
-    // Get a file by ID
-    getFileById(id) {
-      return this.filesById[id] || null;
-    },
-    
-    // Get files by parent ID
-    getFilesByParent(parentId) {
-      return this.filesByParent[parentId] || [];
-    },
-    
-    // Get a file by name and parent
-    getFileByNameAndParent(name, parentId) {
-      const key = `${parentId}:${name}`;
-      return this.filesByNameAndParent[key] || null;
-    },
-    
-    // Mark the cache as stale
-    markStale() {
-      console.log('Marking drive cache as stale');
-      this.isStale = true;
-    },
-    
-    // Clear the entire cache
-    clear() {
-      console.log('Clearing drive cache');
-      this.initialized = false;
-      this.isStale = false;
-      this.filesById = {};
-      this.filesByParent = {};
-      this.filesByNameAndParent = {};
-    },
-    
-    // Fetch files from Drive API and update the cache
-    async fetchFiles({
-      parentId = null,
-      name = null,
-      mimeTypes = [],
-      fields = 'files(id, name, mimeType, parents)',
-      orderBy = null,
-      trashed = false,
-      context = 'fetching files'
-    }) {
-      try {
-        // Build the query parts
-        const queryParts = [];
-        
-        // Add parent folder constraint if provided
-        if (parentId) {
-          queryParts.push(`'${parentId}' in parents`);
-        }
-        
-        // Add name constraint if provided
-        if (name) {
-          queryParts.push(`name='${name}'`);
-        }
-        
-        // Add mime type constraints if provided
-        if (mimeTypes.length > 0) {
-          const mimeTypeQuery = mimeTypes.map(type => `mimeType='${type}'`).join(' or ');
-          queryParts.push(`(${mimeTypeQuery})`);
-        }
-        
-        // Add trashed constraint
-        queryParts.push(`trashed=${trashed}`);
-        
-        // Combine all query parts with AND
-        const query = queryParts.join(' and ');
-        
-        // Build the request parameters
-        const params: any = {
-          q: query,
-          fields,
-          pageSize: 1000
-        };
-        
-        // Add orderBy if provided
-        if (orderBy) {
-          params.orderBy = orderBy;
-        }
-        
-        console.log(`Making Drive API list call for: ${context}`);
-        // Execute the request
-        const { result } = await gapi.client.drive.files.list(params);
-        const files = result.files || [];
-        
-        // Add files to cache
-        files.forEach(file => this.addFile(file));
-        
-        // If this was a folder listing, mark that we've cached this folder's contents
-        if (parentId && files.length > 0) {
-          this.filesByParent[parentId] = files;
-        }
-        
-        return files;
-      } catch (error) {
-        // Check if this is an error that might indicate our cache is out of sync
-        if (error.errorType === DriveErrorType.NOT_FOUND || 
-            error.status === 404 || 
-            (error.message && error.message.toLowerCase().includes('not found'))) {
-          this.markStale();
-          console.warn('Resource not found, marking cache as stale');
-        }
-        
-        handleDriveError(error, context);
-        return [];
-      }
-    }
-  };
-
-  /**
-   * Helper function to lazily get a specific file ID from Drive
-   * @param name File name to look for
-   * @param parentId Parent folder ID
-   * @param forceRefresh Whether to force a cache refresh
-   * @returns The file ID if found, null otherwise
-   */
-  async function getFileId(name, parentId, forceRefresh = false) {
-    if (!parentId) return null;
-    
-    // Check if we have this file in the cache
-    if (!forceRefresh && !driveCache.isStale) {
-      const cachedFile = driveCache.getFileByNameAndParent(name, parentId);
-      if (cachedFile) {
-        return cachedFile.id;
-      }
-    }
-    
-    // If not in cache or cache is stale, fetch from Drive
-    const files = await driveCache.fetchFiles({
-      parentId,
-      name,
-      fields: 'files(id, name, parents)',
-      context: `getting ID for ${name}`
-    });
-    
-    return files.length > 0 ? files[0].id : null;
-  }
-  
-  /**
    * Helper function to lazily get the volume data file ID
    * @param forceRefresh Whether to force a cache refresh
    * @returns The volume data file ID if found, null otherwise
@@ -1249,35 +1042,6 @@
     const id = await getFileId(PROFILES_FILE, readerFolderId, forceRefresh);
     if (id) profilesId = id;
     return id;
-  }
-
-  /**
-   * Function to list files in a folder using our cache
-   * @param parentId The parent folder ID
-   * @param options Additional options
-   * @returns Array of files in the folder
-   */
-  async function listFilesInFolder(parentId, options = {}) {
-    const { forceRefresh = false, mimeTypes = [] } = options;
-    
-    // Check if we have this folder's contents in the cache
-    if (!forceRefresh && !driveCache.isStale) {
-      const cachedFiles = driveCache.getFilesByParent(parentId);
-      if (cachedFiles.length > 0) {
-        // Filter by mime types if specified
-        if (mimeTypes.length > 0) {
-          return cachedFiles.filter(file => mimeTypes.includes(file.mimeType));
-        }
-        return cachedFiles;
-      }
-    }
-    
-    // If not in cache or cache is stale, fetch from Drive
-    return await driveCache.fetchFiles({
-      parentId,
-      mimeTypes,
-      context: `listing files in folder ${parentId}`
-    });
   }
 
   // Function to process a folder recursively
